Redirect URL corrected
If the project were inside a folder, instead of being in the root folder
of the server, the URL built would not contain the path to the folder.
As an example, if it was expected the URL
'http://localhost/MyPersonalWebpage/start_page.php', the URL built would
be 'http://localhost/start_page.php'.